move sliders and share menu into components
