+++ frontend/src/features/event/results/components/share-menu.tsx
